Use annotations as a drop-in replacement for TypesDB
